# scalable-api/src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"CreateUserDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"email",
					"phone"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateUserDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "scalable-api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/users": {
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"properties": {
											"updatedAt": {
												"type": "string",
												"format": "date-time"
											},
											"createdAt": {
												"type": "string",
												"format": "date-time"
											},
											"phone": {
												"type": "string"
											},
											"email": {
												"type": "string"
											},
											"id": {
												"type": "string"
											},
											"name": {
												"type": "string"
											}
										},
										"required": [
											"updatedAt",
											"createdAt",
											"phone",
											"email",
											"id",
											"name"
										],
										"type": "object"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserDto"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "FindAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"properties": {
											"updatedAt": {
												"type": "string",
												"format": "date-time"
											},
											"createdAt": {
												"type": "string",
												"format": "date-time"
											},
											"phone": {
												"type": "string"
											},
											"email": {
												"type": "string"
											},
											"name": {
												"type": "string"
											},
											"id": {
												"type": "string"
											}
										},
										"required": [
											"updatedAt",
											"createdAt",
											"phone",
											"email",
											"name",
											"id"
										],
										"type": "object"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": []
			}
		},
		"/users/{id}": {
			"get": {
				"operationId": "FindOne",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"updatedAt": {
											"type": "string",
											"format": "date-time"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										},
										"phone": {
											"type": "string"
										},
										"email": {
											"type": "string"
										},
										"name": {
											"type": "string"
										},
										"id": {
											"type": "string"
										}
									},
									"required": [
										"updatedAt",
										"createdAt",
										"phone",
										"email",
										"name",
										"id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Get a specific user by ID\r\nTSOA uses the",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"updatedAt": {
											"type": "string",
											"format": "date-time"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										},
										"phone": {
											"type": "string"
										},
										"email": {
											"type": "string"
										},
										"name": {
											"type": "string"
										},
										"id": {
											"type": "string"
										}
									},
									"required": [
										"updatedAt",
										"createdAt",
										"phone",
										"email",
										"name",
										"id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUserDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"updatedAt": {
											"type": "string",
											"format": "date-time"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										},
										"phone": {
											"type": "string"
										},
										"email": {
											"type": "string"
										},
										"id": {
											"type": "string"
										},
										"name": {
											"type": "string"
										}
									},
									"required": [
										"updatedAt",
										"createdAt",
										"phone",
										"email",
										"id",
										"name"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}